+++ experimental/ListView/ListFooter.vue
@@ -26,7 +26,7 @@
   </div>
 </template>
 <script setup>
-import TabButtons from '../TabButtons/TabButtons.vue'
+import TabButtons from '#components/TabButtons/TabButtons.vue'
 import { ref, computed } from 'vue'
 
 const props = defineProps({
