+++ experimental/ListView/stories/Empty.vue
@@ -1,6 +1,7 @@
 <script setup>
 import { h, reactive } from 'vue'
-import { Avatar, ListView } from 'frappe-ui'
+import { Avatar } from 'frappe-ui'
+import { ListView } from 'frappe-ui/experimental'
 
 const cols = reactive([
   {
