+++ experimental/ListView/stories/Examples.vue
@@ -1,7 +1,8 @@
 <script setup>
 import { h, reactive } from 'vue'
 
-import { Avatar, ListView } from 'frappe-ui'
+import { Avatar } from 'frappe-ui'
+import { ListView } from 'frappe-ui/experimental'
 
 const columns = reactive([
   {
